[POSTPONED] Search options list for a value
Make Select get the current value from its available options (based on the supplied value) rather than trying to display the actual passed value.
The value may not have `label` in it resulting in an empty string displayed in Select on UI.
By searching for a value in available options list we will make sure the value is displayed correctly (every option has a label).

This is postponed because of some bugs and the fact that this is not actually needed currently.